tbd6_complete/crm_persister/requirements.txt
confluent-kafka==2.0.2
psycopg2-binary==2.9.9
